refactor: make libraries schema into a new file
